requirements.txt
requests=2.32.4
notion-client=2.4.0
google-generativeai=0.8.5
youtube-transcript-api=1.1.1
